+++ analysis/analysers/AnalyserInterface/src/main/kotlin/de/maibornwolff/codecharta/analysers/analyserinterface/localchanges/LocalChangesResult.kt
@@ -1,0 +1,10 @@
+package de.maibornwolff.codecharta.analysers.analyserinterface.localchanges
+
+data class LocalChangesResult(
+    val changedFiles: Set<String>,
+    val deletedFiles: Set<String>
+) {
+    companion object {
+        val EMPTY = LocalChangesResult(emptySet(), emptySet())
+    }
+}