+++ analysis/analysers/parsers/RawTextParser/src/main/kotlin/de/maibornwolff/codecharta/analysers/parsers/rawtext/RawTextParser.kt
@@ -3,6 +3,9 @@ package de.maibornwolff.codecharta.analysers.parsers.rawtext
 import de.maibornwolff.codecharta.analysers.analyserinterface.AnalyserDialogInterface
 import de.maibornwolff.codecharta.analysers.analyserinterface.AnalyserInterface
 import de.maibornwolff.codecharta.analysers.analyserinterface.CommonAnalyserParameters
+import de.maibornwolff.codecharta.analysers.analyserinterface.basefile.BaseFileResolver
+import de.maibornwolff.codecharta.analysers.analyserinterface.localchanges.LocalChangesDetector
+import de.maibornwolff.codecharta.analysers.analyserinterface.localchanges.LocalChangesResult
 import de.maibornwolff.codecharta.analysers.analyserinterface.util.CommaSeparatedParameterPreprocessor
 import de.maibornwolff.codecharta.analysers.analyserinterface.util.CommaSeparatedStringToListConverter
 import de.maibornwolff.codecharta.model.AttributeDescriptor
@@ -26,7 +29,8 @@ import java.io.PrintStream
 class RawTextParser(
     private val input: InputStream = System.`in`,
     private val output: PrintStream = System.out,
-    private val error: PrintStream = System.err
+    private val error: PrintStream = System.err,
+    private val baseFileHelper: BaseFileResolver = BaseFileResolver()
 ) : AnalyserInterface, AttributeGenerator, CommonAnalyserParameters() {
     @CommandLine.Option(
         names = ["-m", "--metrics"],
@@ -76,13 +80,19 @@ class RawTextParser(
             "Input invalid file for RawTextParser, stopping execution..."
         }
 
+        val validatedInput = inputFile!!
         val useGitignore = !bypassGitignore
-        val effectivePatternsToExclude = determineExclusionPatterns(inputFile!!, useGitignore)
+        val effectivePatternsToExclude = determineExclusionPatterns(validatedInput, useGitignore)
+        val localChangesResult = if (localChanges) {
+            LocalChangesDetector(validatedInput).getLocallyChangedFiles()
+        } else {
+            LocalChangesResult.EMPTY
+        }
 
-        val baseFileNodeMap = loadBaseFileNodes()
+        val baseFileNodeMap = baseFileHelper.loadBaseFileNodes(baseFile)
         val projectMetricsCollector =
             ProjectMetricsCollector(
-                inputFile,
+                validatedInput,
                 effectivePatternsToExclude,
                 fileExtensionsToAnalyse,
                 metricNames,
@@ -90,9 +100,11 @@ class RawTextParser(
                 maxIndentLvl,
                 tabWidth,
                 baseFileNodeMap,
-                useGitignore
+                useGitignore,
+                localChangesResult.changedFiles
             )
         val projectMetrics: ProjectMetrics = projectMetricsCollector.parseProject()
+        addDeletedFileMetrics(projectMetrics, localChangesResult.deletedFiles)
         println()
 
         reportNotFoundFileExtensions(projectMetrics)
@@ -111,6 +123,15 @@ class RawTextParser(
         ProjectSerializer.serializeToFileOrStream(project, outputFile, output, compress)
 
         return null
+    }
+
+    private fun addDeletedFileMetrics(projectMetrics: ProjectMetrics, deletedFiles: Set<String>) {
+        for (filePath in deletedFiles) {
+            projectMetrics.addFileMetrics("/$filePath", FileMetrics())
+        }
+        if (deletedFiles.isNotEmpty()) {
+            Logger.info { "${deletedFiles.size} deleted files included with empty metrics" }
+        }
     }
 
     private fun determineExclusionPatterns(inputFile: File, useGitignore: Boolean): List<String> {